metrics_v2/publisher/requirements.txt
absl-py==0.11.0
google-cloud-pubsub==2.2.0
kubernetes==12.0.1
protobuf==3.14.0
